Python/homework uncertainty (Knowledge-based Systems)/orodaelturrim/requirements.txt
pyqt5
click
antlr4-python3-runtime
multi_key_dict